Got the cons function working and I dont know how

+++ src/edu/osu/cse6341/lispInterpreter/tokenizer/tokens/CloseToken.java
@@ -15,7 +15,7 @@ public class CloseToken implements IToken{
 
 	@Override
 	public int hashCode(){
-		return (int) getTokenKind().ordinal();
+		return getTokenKind().ordinal();
 	}
 
 	@Override
