+++ src/edu/osu/cse6341/lispInterpreter/tokenizer/tokens/EndOfFileToken.java
@@ -15,7 +15,7 @@ public class EndOfFileToken implements IToken{
 	
 	@Override
 	public int hashCode(){
-		return (int) getTokenKind().ordinal();
+		return getTokenKind().ordinal();
 	}
 
 	@Override
